feat: refactor sources, move files
